Move the `client-bundle` module under the `@bayou` namespace.
