+++ local-modules/@bayou/testing-server/README.md
@@ -9,8 +9,8 @@ This module serves multiple related purposes:
 
 A bit of subtlety here is that it is code on the server side which prepares
 the client-side testing code. This module helps with that. In that regard, it
-is similar to the `client-bundle` module, which is a _server_ module responsible
-for manipulating _client_ code.
+is similar to the `@bayou/client-bundle` module, which is a _server_ module
+responsible for manipulating _client_ code.
 
 - - - - - - - - - -
