+++ local-modules/app-setup/package.json
@@ -6,7 +6,7 @@
     "@bayou/api-common": "local",
     "@bayou/api-server": "local",
     "@bayou/app-common": "local",
-    "client-bundle": "local",
+    "@bayou/client-bundle": "local",
     "@bayou/codec": "local",
     "@bayou/deps-ansi-color": "local",
     "@bayou/deps-express": "local",
